backend/pyproject.toml
[project]
name = "schoolos-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic-settings>=2.4",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "orjson>=3.10",
]

[tool.black]
line-length = 100
[tool.ruff]
line-length = 100
